+++ gwcelery/tests/test_tasks_external_triggers.py
@@ -13,8 +13,7 @@ from . import data
 @pytest.mark.parametrize('pipeline, path',
                          [['Fermi', 'fermi_grb_gcn.xml'],
                           ['INTEGRAL', 'integral_grb_gcn.xml'],
-                          ['INTEGRAL_MDC', 'integral_mdc_gcn.xml'],
-                          ['AGILE', 'agile_grb_gcn.xml']])
+                          ['INTEGRAL_MDC', 'integral_mdc_gcn.xml']])
 @patch('gwcelery.tasks.external_skymaps.create_upload_external_skymap.run')
 @patch('gwcelery.tasks.external_skymaps.get_upload_external_skymap.run')
 @patch('gwcelery.tasks.detchar.dqr_json', return_value='dqrjson')
@@ -66,12 +65,10 @@ def test_handle_create_grb_event(mock_create_event,
             ['data_quality'])
     ]
     mock_upload.assert_has_calls(calls, any_order=True)
-    gcn_type_dict = {'Fermi': 115, 'INTEGRAL': 53, 'INTEGRAL_MDC': 53,
-                     'AGILE': 105}
+    gcn_type_dict = {'Fermi': 115, 'INTEGRAL': 53, 'INTEGRAL_MDC': 53}
     time_dict = {'Fermi': '2018-05-24T18:35:45',
                  'INTEGRAL': '2017-02-03T19:00:05',
-                 'INTEGRAL_MDC': '2023-04-04T06:31:24',
-                 'AGILE': '2019-03-19T19:40:49'}
+                 'INTEGRAL_MDC': '2023-04-04T06:31:24'}
     mock_create_upload_external_skymap.assert_called_once_with(
         {'graceid': 'E1',
          'gpstime': 1,